requirements.txt
google-generativeai==0.8.3
python-dotenv==1.0.1
secure-smtplib==0.1.1
pypdf==5.3.0
